Task_2/requirements.txt
tabula-py==2.7.0
tabulate==0.9.0
numpy==1.23.5
imutils==0.5.4
pytesseract==0.3.10
Pillow==9.4.0
mmdet==3.0.0
mmcv==2.0.0
torch==2.0.0
opencv-python==4.7.0.72